+++ packages/alfa-iterable/src/iterable.ts
@@ -22,13 +22,13 @@ export namespace Iterable {
     return isObject(value) && Symbol.iterator in value;
   }
 
+  export function* empty<T>(): Iterable<T> {}
+
   export function* from<T>(arrayLike: ArrayLike<T>): Iterable<T> {
     for (let i = 0, n = arrayLike.length; i < n; i++) {
       yield arrayLike[i];
     }
   }
-
-  export function* empty<T>(): Iterable<T> {}
 
   export function size<T>(iterable: Iterable<T>): number {
     return reduce(iterable, (size) => size + 1, 0);
@@ -283,10 +283,105 @@ export namespace Iterable {
     return index < 0 ? false : !isEmpty(skip(iterable, index));
   }
 
-  export function* concat<T>(...iterables: Array<Iterable<T>>): Iterable<T> {
+  export function* set<T>(
+    iterable: Iterable<T>,
+    index: number,
+    value: T
+  ): Iterable<T> {
+    const it = iterator(iterable);
+
+    while (index-- > 0) {
+      const next = it.next();
+
+      if (next.done === true) {
+        return;
+      }
+
+      yield next.value;
+    }
+
+    const next = it.next();
+
+    if (next.done === true) {
+      return;
+    }
+
+    yield value;
+
+    while (true) {
+      const next = it.next();
+
+      if (next.done === true) {
+        return;
+      }
+
+      yield next.value;
+    }
+  }
+
+  export function* insert<T>(
+    iterable: Iterable<T>,
+    index: number,
+    value: T
+  ): Iterable<T> {
+    const it = iterator(iterable);
+
+    while (index-- > 0) {
+      const next = it.next();
+
+      if (next.done === true) {
+        return;
+      }
+
+      yield next.value;
+    }
+
+    yield value;
+
+    while (true) {
+      const next = it.next();
+
+      if (next.done === true) {
+        return;
+      }
+
+      yield next.value;
+    }
+  }
+
+  export function* append<T>(iterable: Iterable<T>, value: T): Iterable<T> {
+    yield* iterable;
+    yield value;
+  }
+
+  export function* prepend<T>(iterable: Iterable<T>, value: T): Iterable<T> {
+    yield value;
+    yield* iterable;
+  }
+
+  export function* concat<T>(
+    iterable: Iterable<T>,
+    ...iterables: Array<Iterable<T>>
+  ): Iterable<T> {
+    yield* iterable;
+
     for (const iterable of iterables) {
       yield* iterable;
     }
+  }
+
+  export function subtract<T>(
+    iterable: Iterable<T>,
+    ...iterables: Array<Iterable<T>>
+  ): Iterable<T> {
+    return reject(iterable, (value) => includes(flatten(iterables), value));
+  }
+
+  export function intersect<T>(
+    iterable: Iterable<T>,
+    ...iterables: Array<Iterable<T>>
+  ): Iterable<T> {
+    return filter(iterable, (value) => includes(flatten(iterables), value));
   }
 
   export function first<T>(iterable: Iterable<T>): Option<T> {
@@ -308,10 +403,10 @@ export namespace Iterable {
   }
 
   export function* take<T>(iterable: Iterable<T>, count: number): Iterable<T> {
-    const iterator = iterable[Symbol.iterator]();
+    const it = iterator(iterable);
 
     while (count-- > 0) {
-      const next = iterator.next();
+      const next = it.next();
 
       if (next.done === true) {
         return;
@@ -393,10 +488,10 @@ export namespace Iterable {
   }
 
   export function* skip<T>(iterable: Iterable<T>, count: number): Iterable<T> {
-    const iterator = iterable[Symbol.iterator]();
+    const it = iterator(iterable);
 
     while (count-- > 0) {
-      const next = iterator.next();
+      const next = it.next();
 
       if (next.done === true) {
         return;
@@ -404,7 +499,7 @@ export namespace Iterable {
     }
 
     while (true) {
-      const next = iterator.next();
+      const next = it.next();
 
       if (next.done === true) {
         return;
@@ -442,12 +537,12 @@ export namespace Iterable {
     iterable: Iterable<T>,
     count: number = 1
   ): Iterable<T> {
-    const iterator = iterable[Symbol.iterator]();
+    const it = iterator(iterable);
 
     const first: Array<T> = [];
 
     while (count-- > 0) {
-      const next = iterator.next();
+      const next = it.next();
 
       if (next.done === true) {
         return;
@@ -457,7 +552,7 @@ export namespace Iterable {
     }
 
     while (true) {
-      const next = iterator.next();
+      const next = it.next();
 
       if (next.done === true) {
         return;
@@ -545,20 +640,20 @@ export namespace Iterable {
   }
 
   export function join<T>(iterable: Iterable<T>, separator: string): string {
-    const iterator = iterable[Symbol.iterator]();
+    const it = iterator(iterable);
 
-    let next = iterator.next();
+    let next = it.next();
 
     if (next.done === true) {
       return "";
     }
 
     let result = `${next.value}`;
-    next = iterator.next();
+    next = it.next();
 
     while (next.done !== true) {
       result += `${separator}${next.value}`;
-      next = iterator.next();
+      next = it.next();
     }
 
     return result;
@@ -589,12 +684,12 @@ export namespace Iterable {
     b: Iterable<T>,
     comparer: Comparer<T>
   ): Comparison {
-    const iteratorA = a[Symbol.iterator]();
-    const iteratorB = b[Symbol.iterator]();
+    const itA = iterator(a);
+    const itB = iterator(b);
 
     while (true) {
-      const a = iteratorA.next();
-      const b = iteratorB.next();
+      const a = itA.next();
+      const b = itB.next();
 
       if (a.done === true) {
         return b.done === true ? Comparison.Equal : Comparison.Less;
@@ -613,12 +708,12 @@ export namespace Iterable {
   }
 
   export function equals<T>(a: Iterable<T>, b: Iterable<T>): boolean {
-    const iteratorA = a[Symbol.iterator]();
-    const iteratorB = b[Symbol.iterator]();
+    const itA = iterator(a);
+    const itB = iterator(b);
 
     while (true) {
-      const a = iteratorA.next();
-      const b = iteratorB.next();
+      const a = itA.next();
+      const b = itB.next();
 
       if (a.done === true) {
         return b.done === true;
@@ -641,18 +736,8 @@ export namespace Iterable {
     Hash.writeUint32(hash, size);
   }
 
-  export function subtract<T>(
-    left: Iterable<T>,
-    right: Iterable<T>
-  ): Iterable<T> {
-    return filter(left, (left) => !includes(right, left));
-  }
-
-  export function intersect<T>(
-    left: Iterable<T>,
-    right: Iterable<T>
-  ): Iterable<T> {
-    return filter(left, (left) => includes(right, left));
+  export function iterator<T>(iterable: Iterable<T>): Iterator<T> {
+    return iterable[Symbol.iterator]();
   }
 
   export function groupBy<T, K>(